+++ pages/compras.php
@@ -118,6 +118,62 @@ desired effect
 |---------------------------------------------------------|
 -->
 <body class="hold-transition skin-blue sidebar-mini">
+<script>
+    // Populo gerencias
+    function fn_popular_comentarios(id_compra){
+      // Busco el servicio
+      $.ajax({
+          type: 'POST',
+          url: './helpers/getAsyncDataFromDB.php',
+          data: { query: "select c.*, p.nombre, p.apellido FROM adm_compras_comments as c INNER JOIN persona as p ON c.id_persona = p.id_persona WHERE c.id_compra = " + id_compra + " AND c.borrado = 0 ORDER BY c.id DESC" },
+          dataType: 'json',
+          success: function(json) {
+                $('#chat-box').empty();
+                if ("data" in json == true) {
+                    $.each(json.data, function(i, d) {
+                        $('#chat-box').append('<div class="item">')
+                        $('#chat-box').append('<i class="fa fa-info-circle" style="font-size:16px;"></i>');
+                        $('#chat-box').append('<a href="#" class="name"><small class="text-muted pull-right"><i class="fa fa-clock-o"></i> ' + d.fecha + '</small>&nbsp;' + d.apellido + ', ' + d.nombre + '</a>');
+                        $('#chat-box').append('<p>' + d.comentario + '</p>'); 
+                        $('#chat-box').append('</div>'); 
+                    });
+                }
+          },
+          error: function(xhr, status, error) {
+            $('#chat-box').empty();
+            alert(xhr.responseText, error);
+          }
+      });
+    }
+
+
+    function addcomment() { 
+      // console.log($('#compra-selected-id').val());
+        // Ejecuto
+        let id_compra = $('#compra-selected-id').val();
+        let comentario = $('#popover-comment').val();
+        let usuario = $('#compra-id-persona').val();
+        console.log(id_compra, '', comentario, '', usuario);
+        $.ajax({
+            type: 'POST',
+            url: './helpers/abmcompracommentdb.php',
+            data: {
+                operacion: 'A',
+                id_compra: id_compra,
+                comentario: comentario,
+                id_persona: usuario
+            },
+            dataType: 'json',
+            success: function(json) {
+                $('#popover-add-comment').popover('hide'); 
+                fn_popular_comentarios(id_compra);
+            },
+            error: function(xhr, status, error) {
+                alert(xhr.responseText, error);
+            }
+        });      
+    };
+</script>
 <div class="wrapper">
 
   <!-- Main Header -->
@@ -132,7 +188,35 @@ desired effect
   <div class="content-wrapper">
     <!-- Content Header (Page header) -->
     <section class="content-header">
-            <h1>Gestión de Compras</h1>
+      <!-- INDICADORES -->
+      <div class="row">
+        <!-- /.box-header -->
+          <div class="col-xs-12 col-md-8"><h1>Gestión de Compras</h1></div>
+          <div class="col-xs-6 col-md-1 text-center">
+            <input type="text" class="knob" value="30" data-width="60" data-height="60" data-fgColor="#3c8dbc">
+            <div class="knob-label">New Visitors</div>
+          </div>
+          <!-- ./col -->
+          <div class="col-xs-6 col-md-1 text-center">
+            <input type="text" class="knob" value="70" data-width="60" data-height="60" data-fgColor="#f56954">
+
+            <div class="knob-label">Bounce Rate</div>
+          </div>
+          <!-- ./col -->
+          <div class="col-xs-6 col-md-1 text-center">
+            <input type="text" class="knob" value="-80" data-min="-150" data-max="150" data-width="60" data-height="60" data-fgColor="#00a65a">
+
+            <div class="knob-label">Server Load</div>
+          </div>
+          <!-- ./col -->
+          <div class="col-xs-6 col-md-1 text-center">
+            <input type="text" class="knob" value="40" data-width="60" data-height="60" data-fgColor="#00c0ef">
+
+            <div class="knob-label">Disk Space</div>
+            <!-- /.row -->
+          </div>
+        <!-- /.box-body -->
+      </div>             
     </section>
     <!-- Main content -->
     <section class="content container-fluid">
@@ -140,41 +224,17 @@ desired effect
     | Your Page Content Here |
     -------------------------->
 	    <section class="content">
-            <!-- INDICADORES -->
-            <div class="row">
-            <!-- /.box-header -->
-                <div class="col-xs-12 col-md-8 text-center"></div>
-                <div class="col-xs-6 col-md-1 text-center">
-                  <input type="text" class="knob" value="30" data-width="90" data-height="90" data-fgColor="#3c8dbc">
-
-                  <div class="knob-label">New Visitors</div>
-                </div>
-                <!-- ./col -->
-                <div class="col-xs-6 col-md-1 text-center">
-                  <input type="text" class="knob" value="70" data-width="90" data-height="90" data-fgColor="#f56954">
-
-                  <div class="knob-label">Bounce Rate</div>
-                </div>
-                <!-- ./col -->
-                <div class="col-xs-6 col-md-1 text-center">
-                  <input type="text" class="knob" value="-80" data-min="-150" data-max="150" data-width="90" data-height="90" data-fgColor="#00a65a">
-
-                  <div class="knob-label">Server Load</div>
-                </div>
-                <!-- ./col -->
-                <div class="col-xs-6 col-md-1 text-center">
-                  <input type="text" class="knob" value="40" data-width="90" data-height="90" data-fgColor="#00c0ef">
-
-                  <div class="knob-label">Disk Space</div>
-              <!-- /.row -->
-            </div>
-            <!-- /.box-body -->
-          </div>         
+        
+        <input type="hidden" class="form-control" name="id_persona" id='compra-id-persona' value="<?=$id_rowp ?>">
             <!-- CONTENIDO -->
             <div class="row">
                 <div class="col-12">
                     <!-- Custom Tabs -->
                     <div class="nav-tabs-custom">
+                        <!-- BOTON -->
+                        <div class="pull-right" style="margin: 10px;">
+                          <button id="modal-abm-compra-btn-alta" type="button" class="btn btbn-block btn-primary btn-sm">Nueva Compra</button>
+                        </div>
                         <ul class="nav nav-tabs">
                             <li class="active"><a href="#tab_1" data-toggle="tab">En proceso</a></li>
                             <li><a href="#tab_2" data-toggle="tab">Adjudicadas</a></li>
@@ -183,8 +243,9 @@ desired effect
                             <!-- En proceso -->
                             <div class="tab-pane active" id="tab_1">
                                 <div class="row">
+                                    <input type="hidden" class="form-control" name="id" id='compra-selected-id' >
                                     <!-- COMPRAS -->
-                                    <div class="col-md-9">
+                                    <div class="col-md-10">
                                         <div class="box">
                                             <div class="box-body">
                                                 <table id="tbEnProceso" class="display w-auto" witdh="70%">
@@ -230,32 +291,45 @@ desired effect
                                                                 echo '<td>'. $row['moneda'] .' ' . $row['pre_monto'] . '</td>';
                                                                 echo '<td>'. $row['cur_step_desc'] .'</td>';
                                                                 echo '<td>'. $row['next_step_desc'] .'</td>';
-                                                                echo '<td><a data-id="'.$row['id'].'" title="ver más info" class="ver-itemDialog btn btn-sm"><i class="glyphicon glyphicon-eye-open"></i></a></td>';
+                                                                echo '<td align="center"><a data-id="'.$row['id'].'" title="ver más info" class="ver-itemDialog btn btn-sm"><i class="glyphicon glyphicon-eye-open"></i></a>';
+                                                                echo '<a data-id="'.$row['id'].'" title="editar" class="ver-itemDialog btn btn-sm"><i class="glyphicon glyphicon-edit" style="color: red;"></i></a></td>';
                                                                 echo '</tr>';
                                                             }
                                                         ?>
                                                     </tbody>
                                                 </table>
+                                                <!-- MODAL ADD PERSONA -->
+                                                <?php
+                                                    include_once('./modals/abmcompra.php');
+                                                ?>                                                
                                             </div>
                                             <!-- /.box-body -->
                                         </div>  
                                     </div>
                                     <!-- COMMENTS -->
-                                    <div class="col-md-3">
+                                    <div class="col-md-2">
                                         <div class="box box-success">
-                                            <div class="box-header">
+                                            <!-- <div class="pull-right"> -->
+                                              <!-- </div> -->
+                                              <div class="box-header">
                                                 <i class="fa fa-comments-o"></i>
-                                                <h3 class="box-title">Comentarios</h3>
+                                                <h3 class="box-title">Comentarios&nbsp;&nbsp;&nbsp;
+                                                    <a id="popover-add-comment" style="color: green; font-size:20px;" title="Agregar commentario" 
+                                                      tabindex="0" data-html="true" 
+                                                      data-Title="Agregar comentario" 
+                                                      data-container="body" 
+                                                      data-toggle="popover" 
+                                                      data-placement="left" 
+                                                      data-trigger="click"
+                                                      data-content='<div class="input-group"><input id="popover-comment" class="form-control">
+                                                      <div class="input-group-btn"><button id="popup-comment-submit" type="button" class="btn btn-success" onclick="addcomment();"><i class="fa fa-plus"></i></button></div>
+                                                      </div>'><i id="popover-add-comment-icon" class="fa fa-plus-circle"></i>
+                                                    </a>                                              
+
+                                                </h3>
                                             </div>
                                             <div class="box-body chat" id="chat-box">
                                             <div class="box-footer">
-                                            <div class="input-group">
-                                                <input class="form-control" placeholder="Type message...">
-
-                                                <div class="input-group-btn">
-                                                <button type="button" class="btn btn-success"><i class="fa fa-plus"></i></button>
-                                                </div>
-                                            </div>
                                             </div>
                                         </div>                                                
                                     </div>                                        
@@ -282,6 +356,8 @@ desired effect
 
 <!-- jQuery 3 -->
 <script src="../bower_components/jquery/dist/jquery.min.js"></script>
+<!-- Popper -->
+<script src="../bower_components/popper/popper.min.js"></script>
 <!-- Bootstrap 3.3.7 -->
 <script src="../bower_components/bootstrap/dist/js/bootstrap.min.js"></script>
 <!-- DataTables -->
@@ -296,35 +372,12 @@ desired effect
 <!-- jQuery Knob Chart -->
 <script src="../bower_components/jquery-knob/dist/jquery.knob.min.js"></script>
 
+<!-- MODALES -->
+<script src="./modals/abmcompra.js"></script>
+
 <script>
   $(function () {
 
-    // Populo gerencias
-    function fn_popular_comentarios(id_compra){
-      // Busco el servicio
-      $.ajax({
-          type: 'POST',
-          url: './helpers/getAsyncDataFromDB.php',
-          data: { query: "select c.*, p.nombre, p.apellido FROM adm_compras_comments as c INNER JOIN persona as p ON c.id_persona = p.id_persona WHERE c.id_compra = " + id_compra + " AND c.borrado = 0 ORDER BY c.id DESC" },
-          dataType: 'json',
-          success: function(json) {
-                $('#chat-box').empty();
-                if ("data" in json == true) {
-                    $.each(json.data, function(i, d) {
-                        $('#chat-box').append('<div class="item">')
-                        $('#chat-box').append('<i class="fa fa-info-circle" style="font-size:16px;"></i>');
-                        $('#chat-box').append('<a href="#" class="name"><small class="text-muted pull-right"><i class="fa fa-clock-o"></i> ' + d.fecha + '</small>&nbsp;' + d.apellido + ', ' + d.nombre + '</a>');
-                        $('#chat-box').append('<p>' + d.comentario + '</p>'); 
-                        $('#chat-box').append('</div>'); 
-                    });
-                }
-          },
-          error: function(xhr, status, error) {
-            $('#chat-box').empty();
-            alert(xhr.responseText, error);
-          }
-      });
-    }
 
 
     $('#tbEnProceso').DataTable({
@@ -337,15 +390,31 @@ desired effect
       'autoWidth'   : true,
     });
 
-    $('#tbEnProceso').on('click', 'tr', function(event){
+    $('#tbEnProceso tbody').on('click', 'tr', function(event){
+     
         let tb = $('#tbEnProceso').dataTable();
         let datarow = tb.fnGetData(this);
         let id = datarow[0];
+
+        // efecto visual de seleccionar la fila
+        if ( $(this).hasClass('selected') ) {
+            $(this).removeClass('selected');
+        } else {
+          tb.$('tr.selected').removeClass('selected');
+          $(this).addClass('selected');
+        }
+        $("#popover-add-comment-icon").css('color: green;');
+        // seteo el id de la fila seleccionada para que lo use el commentario
+        $('#compra-selected-id').val(id);
         fn_popular_comentarios(id);
-    });
+        $("#popover-add-comment").popover('enable');
+      });
+      
+    $("#popover-add-comment").popover('disable');
+    
+    $("#popover-add-comment-icon").css('color: gray;');
 
     /* jQueryKnob */
-
     $(".knob").knob({
       /*change : function (value) {
        //console.log("change : " + value);
@@ -399,6 +468,13 @@ desired effect
           return false;
         }
       }
+    });
+
+    // Popper
+    $('[data-toggle="popover"]').popover();
+    $('#popover-add-comment').on('shown.bs.popover', function () {
+      $('#popover-comment').val('');
+      $('#popover-comment').focus();
     });
 });
 </script>